Replacing `printf` by ES6 template strings
This adds a `.babelrc` file that enables ES6 transpilation, allowing us to take advantage of handy modern JS features that allow for more succinct code.

These features include template strings, which are a convenient alternative to the `String.prototype.printf` helper method defined in `site.js`. This branch eliminates `printf` and replaces it in all instances with template strings.