ODPM-68 - roll back pull request 80 (agency slugs)
This avoids the need to

1. figure out how to manually create the unique index for slugs
   in copy.sql
2. axe lookup by id (or any other changes to finalize slugs)
3. implement a mechanism to rebuild slugs when reloading MD data

A couple of items could conceivably get worse by this rollback:

The migrate commands in .travis.yml might be useful independent
of agency slugs.  We'll address that as appropriate if we still
need it.

At least one of the NC top five agency links is broken.  Another
ticket, ODPM-51, exists to correct that.