ODPM-80 - API unit tests for NC and MD
This takes the existing branch for NC API tests, gets it
in sync with the current code, updates some of the tests to
check more details, implements a consistent set of tests for
MD, then adds a test of contraband-hit-rate for both.

The NC use-of-force endpoint (not relevant for MD) is not
tested.