ODPM-130: New fix for anchor link scrolling
IE has some weird behavior related to overflow hiding and scrolling. Well, maybe it's not weird: IE refuses to scroll to the top of an element when that top is not visible. It will only scroll to the highest visible point. All other browsers behave differently.

This new implementation of the scrolling fix dispenses with the padding trickery that made overflow-hiding necessary in the first place. Instead it uses an arealess floating `a` element positioned above the target and containing its ID. This has the further advantage of removing the need for extra JS.

To verify:

- Fire up IE.
- Go to an agency detail page.
- Click one of the links in the header (e.g. "Departmental Search Rate"). Verify that it scrolls in such a way that the headline text is visible.
- For good measure, do the same with a mobile-width window.