Adding home page items and navigation
Adds Illinois image to the home page and adjusts HTML and styles to accommodate the new image and button.

Adjusts the navigation template to take into account Illinois URLs. *Note: the implementation of the navigation is increasingly non-scalable and will have to be fixed in future iterations (or later in this iteration).*

Augments middleware to insert 'il' into request.state as appropriate.