ODPM-164: prepending chart select dropdowns
On charts that are preceded by help text that says something about using the chart's dropdown menu (e.g. the Departmental Stop Count chart), we'd like the dropdown to be closer to that text, i.e. above the chart and not after it.

This changes the `appendTo` call for dropdowns to `prependTo`, and it adjusts the appearance of those dropdowns' container elements to make them a bit more consistent and aesthetically acceptable.