Fixing jsdom deps
It turns out that we need to include a 9.x version of jsdom for tests to run.

Once this is there, we can omit `mocha-jsdom` altogether, which seems like a good idea: it's not deprecated, but its page in NPM encourages you to not use it ("consider jsdom-global, a simpler alternative that also works outside of Mocha"), not a great sign for its future.